stm32f4xx: clocks: refactoring: let peripherals access the clocks controller
Introduce a ClocksCtrl trait to let peripheral clocks access clocks
controller features without the need of the ChipSpecific bind.

Note: this moves the clocks struct initialization into the board crates
since it has to be already instantiated when passing its reference to
the peripherals. This commit also removes the rcc reference passing to
peripherals since it is now accessed from the ClocksCtrl trait object.