stm32f4xx: clocks: refactoring: append mhz suffix to freq vars and methods
The clocks module uses frequencies in MHz base unit. Also, peripherals
needs the system clock frequency in Hz base unit. Thus, for clarity,
make it explicit when MHz base unit is used and consider Hz as base unit
when not specified. This seems to be the general policy around the code
base, for instance the kernel::hil::time::Frequency trait.