stm32f4xx: change HCLK frequency getter method for peripheral clocks
Peripheral clocks used to rely on the rcc module to retrieve the HCLK
frequency (mistakenly called SYSCLK which should not consider AHB
prescaler). This had the limitation that the rcc module was not aware of
the HSE clock settings already managed by the clocks module, meaning
that dynamic frequency operations were not supported with the HSE source
clock in use.

This commit exposes the HCLK frequency calculated by the clocks module
to peripheral clocks, thus extending the support for the HSE source
clock.

Also, the methods to retrieve the PLL and SYSCLK frequency (without AHB
division) had been moved from the rcc module to the pll and clocks
modules respectively. They are not necessary anymore since PLL and
SYSCLK frequencies are being retrieved from cached variables, but kept
anyway for future use.